+++ src/main/java/de/cowtipper/cowlection/chesttracker/LowestBinsCache.java
@@ -1,0 +1,10 @@
+package de.cowtipper.cowlection.chesttracker;
+
+import java.util.HashMap;
+
+public class LowestBinsCache extends HashMap<String, Integer> {
+
+    public boolean hasData() {
+        return size() > 0;
+    }
+}